add more infos on how to profile the javascript #10944
Summary:
As said in #10944, there's not yet some good infos on how to profile the javascript.

I'm adding a mention to two ways of doing it (`react-addons-perf` and chrome profiler), feel free to correct me on this.

I almost added an example for `react-addons-perf` but I'm not sure what's the correct way to use. Here's the way I use it:

```javascript
import Perf from 'react-addons-perf';

....
    componentDidMount() {
        console.log('start perf tracking');
        Perf.start();
        setTimeout(() => {
            console.log('stop perf tracking');
            Perf.stop();
            Perf.printExclusive();
        }, 10000);
   }
...
```
Closes https://github.com/facebook/react-native/pull/10974

Differential Revision: D4221630

Pulled By: mkonicek

fbshipit-source-id: 918f837b9c7169c3dd835e653c78159b801fb946