Fixing bad link to RN documentation.
Summary:
The link referencing the root of RN documentation was leading to a 404 error. As the docs don't seem to have a true "index" page, it makes sense that it goes to the top level entry in the docs, the "Getting Started Section". The alternative solution may be not linking that phrase at all, this seems helpful though.

No changes in this pull request require any sort of unit/AB testing, the link directing to the correct place should be apparent.

[DOCS] [BUGFIX] [README.md] - Fixing doc resource link.
Closes https://github.com/facebook/react-native/pull/17048

Differential Revision: D6464455

Pulled By: hramos

fbshipit-source-id: 710dc882672d762478910538341b57cb1bb129bb